sparse_format.h: Fix incorrect chunk_sz field
chunk_sz in chunk_header was incorrectly defined
as a 16 bits field.

reference: http://git.omapzoom.org/?p=platform/
system/core.git;a=blob;f=libsparse/sparse_format.h;
h=c41f12a545c9952d2ab803779af89916ea00618e;hb=refs/
heads/master

This patch corrects that by setting it to 32 bits to
reflect the structure in libsparse.

Issue found during tests of >512MB blocks flash,
./fastbooot -S <size>[K|M|G]

+++ include/common/sparse_format.h
@@ -51,7 +51,7 @@ typedef struct chunk_header {
 	u16	chunk_type;	/* 0xCAC1 -> raw; 0xCAC2 -> fill;
 				0xCAC3 -> don't care */
 	u16	reserved1;
-	u16	chunk_sz;	/* in blocks in output image */
+	u32	chunk_sz;	/* in blocks in output image */
 	u32	total_sz;	/* in bytes of chunk input file
 				including chunk header and data */
 } chunk_header_t;
